requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-request==1.6.2
exceptiongroup==1.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
outcome==1.2.0
packaging==23.0
pluggy==1.0.0
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.8.2
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webencodings==0.5.1
wsproto==1.2.0
